Working LDAP connection
Set up ldapdb to read from our ldap database. Works over internet
too(this means you can connect to LDAP from anywhere, provided you
havebind with a CSUA account).

Because we are now working with live data, we should probably set
upautomated backups of the database, or we may risk losing important
data.

We will also need to renew the cert at some point, we'll have to figure
out how to automate that, too.

Also have to fill in the user info to access the database.

Also have to fix the admin view for the groups, it's a bit messed up.

Also, ldapdb does not support migrations. We would have to make schema
changes directly to the openldap database